requirements/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
adbc-driver-manager==1.6.0
    # via adbc-driver-sqlite
adbc-driver-sqlite==1.6.0
    # via -r requirements.in
click==8.1.8
    # via typer
importlib-resources==6.5.2
    # via adbc-driver-sqlite
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
polars==1.29.0
    # via -r requirements.in
pyarrow==20.0.0
    # via -r requirements.in
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via -r requirements.in
rich==14.0.0
    # via typer
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
typer==0.15.4
    # via -r requirements.in
typing-extensions==4.13.2
    # via
    #   adbc-driver-manager
    #   typer
